requirements.txt
streamlit
opencv-python
numpy
matplotlib
scipy
pillow
pandas
fpdf
